+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Statue.java
@@ -24,6 +24,7 @@ package com.shatteredpixel.shatteredpixeldungeon.actors.mobs;
 import com.shatteredpixel.shatteredpixeldungeon.Dungeon;
 import com.shatteredpixel.shatteredpixeldungeon.actors.Char;
 import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.Buff;
+import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.SwarmGen;
 import com.shatteredpixel.shatteredpixeldungeon.items.Generator;
 import com.shatteredpixel.shatteredpixeldungeon.items.trinkets.RatSkull;
 import com.shatteredpixel.shatteredpixeldungeon.items.weapon.Weapon;
@@ -156,7 +157,8 @@ public class Statue extends Mob {
 	@Override
 	public void die( Object cause ) {
 		//silently duplicated in this case, so don't drop original weapon
-		if (buff(Corrupting.CorruptingTracker.class) == null) {
+		if (buff(Corrupting.CorruptingTracker.class) == null
+				&& !SwarmGen.isClone(this)) {
 			weapon.identify(false);
 			Dungeon.level.drop(weapon, pos).sprite.drop();
 		}
